requirements.txt
pandas
numpy
google-cloud-bigquery
google-auth
gspread
pyyaml
python-dotenv
apache-airflow